remove: Section -> Line에 종속
