+++ src/main/java/subway/section/repository/SectionRepository.java
@@ -1,6 +1,7 @@
 package subway.section.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.line.repository.Section;
 
 public interface SectionRepository extends JpaRepository<Section, Long> {
 }
