+++ src/test/java/subway/utils/LineUtil.java
@@ -9,6 +9,7 @@ import org.springframework.http.MediaType;
 import java.util.HashMap;
 import java.util.Map;
 
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static subway.utils.AssertUtil.*;
 
 public class LineUtil {
@@ -86,5 +87,29 @@ public class LineUtil {
                 .extract();
 
         assertSuccessNoContent(response);
+    }
+
+    public static ExtractableResponse<Response> addSectionResponse(Long lineId, Long downStationId, Long newStationId, int distance) {
+        return RestAssured
+                .given().log().all().body(getAddParam(downStationId, newStationId, distance)).contentType(APPLICATION_JSON_VALUE)
+                .when().post("/lines/{id}/sections", lineId)
+                .then().log().all()
+                .extract();
+    }
+
+    private static Map<String, Object> getAddParam(Long downStationId, Long newStationId, int distance) {
+        Map<String, Object> param = new HashMap<>();
+        param.put("upStationId", downStationId);
+        param.put("downStationId", newStationId);
+        param.put("distance", distance);
+        return param;
+    }
+
+    public static ExtractableResponse<Response> deleteSectionResponse(Long lineId, Long stationId) {
+        return RestAssured
+                .given().log().all().params("stationId", stationId)
+                .when().delete("/lines/{id}/sections", lineId)
+                .then().log().all()
+                .extract();
     }
 }
